noticefile qnafile

+++ src/main/java/com/jh/s1/board/BoardDTO.java
@@ -3,48 +3,59 @@ package com.jh.s1.board;
 import java.sql.Date;
 
 public class BoardDTO {
-	private Integer num;
+	private Long num;
 	private String title;
 	private String contents;
 	private String writer;
 	private Date regDate;
 	private Integer hit;
-	public Integer getNum() {
+
+	public Long getNum() {
 		return num;
 	}
-	public void setNum(Integer num) {
+
+	public void setNum(Long num) {
 		this.num = num;
 	}
+
 	public String getTitle() {
 		return title;
 	}
+
 	public void setTitle(String title) {
 		this.title = title;
 	}
+
 	public String getContents() {
 		return contents;
 	}
+
 	public void setContents(String contents) {
 		this.contents = contents;
 	}
+
 	public String getWriter() {
 		return writer;
 	}
+
 	public void setWriter(String writer) {
 		this.writer = writer;
 	}
+
 	public Date getRegDate() {
 		return regDate;
 	}
+
 	public void setRegDate(Date regDate) {
 		this.regDate = regDate;
 	}
+
 	public Integer getHit() {
 		return hit;
 	}
+
 	public void setHit(Integer hit) {
 		this.hit = hit;
 	}
-	
-	
+
 }
